Unify real and dry run report generation in SummariseCopyLambda.
