+++ packages/steps-s3-copy/lambda/summarise-copy-lambda/summarise-copy-lambda.ts
@@ -7,6 +7,7 @@ import { basename } from "path/posix";
 import { stringify } from "csv-stringify/sync";
 import { dirname } from "path/posix";
 import { createHtmlReport } from "./create-html-report.ts";
+import { createDryRunHtmlReport, FileSummary } from "./create-dryrun-report.ts";
 
 interface InvokeEvent {
   rcloneResultsLarge: {
@@ -32,6 +33,7 @@ interface InvokeEvent {
   copyInstructionsKey: string;
   includeCopyReport?: boolean;
   retainCopyReport?: boolean;
+  dryRun?: boolean;
   inputCopySets: {
     small: { bucket: string; key: string };
     large: { bucket: string; key: string };
@@ -101,6 +103,41 @@ async function readCostsFromJsonl(
   return costs;
 }
 
+export async function readFileSummariesFromJsonl(
+  client: S3Client,
+  bucket: string,
+  key: string,
+): Promise<FileSummary[]> {
+  const getCommand = new GetObjectCommand({ Bucket: bucket, Key: key });
+  const result = await client.send(getCommand);
+
+  if (!result.Body) return [];
+
+  const content = await result.Body.transformToString();
+  const lines = content
+    .trim()
+    .split(/\r?\n/)
+    .filter((l) => l.trim());
+
+  const out: FileSummary[] = [];
+
+  for (const line of lines) {
+    const obj = JSON.parse(line);
+    const cost = obj.costEstimate;
+    if (cost) {
+      out.push({
+        name: obj.sourceKey, // adjust as needed
+        size: obj.size || 0,
+        s3CrossRegionReadWriteCostAUD: cost.s3CrossRegionReadWriteCostAUD || 0,
+        coldStorageRetrievalCostAUD: cost.coldStorageRetrievalCostAUD || 0,
+        computeCostAUD: cost.computeCostAUD || 0,
+        // ... other fields as needed
+      });
+    }
+  }
+  return out;
+}
+
 /**
  * A handler that process the Steps CSV result (which should have a bunch
  * of rclone stats) - and converts that into a CSV report for the
@@ -135,272 +172,45 @@ export async function handler(event: InvokeEvent) {
     event.inputCopySets.largeThaw.key,
   );
 
-  // Each Distributed Map we run (Large, Small, NeedThawSmall, NeedThawLarge) writes a
-  // `manifest.json` to S3. The manifest is generated by Step Functions and summarises the
-  // per-item result files for that map run (grouped into PENDING / FAILED / SUCCEEDED).
-  //
-  // Example:
-  // {"DestinationBucket":"elsa-data-tmp",
-  // "MapRunArn":"arn:aws:states:ap-southeast-2:12345678:mapRun:CopyOutStateMachineABCD/4474d22f-4056-30e3-978c-027016edac90:0c17ffd6-e8ad-44c0-a65b-a8b721007241",
-  // "ResultFiles":{
-  //     "FAILED":[],
-  //     "PENDING":[],
-  //     "SUCCEEDED":[{"Key":"copy-out-test-working/a6faea86c066cd90/1-objects-to-copy.tsv/0c17ffd6-e8ad-44c0-a65b-a8b721007241/SUCCEEDED_0.json",
-  //                   "Size":2887}]}}
-
-  // We read and merge the manifests from all four maps to get the complete set of results.
-
-  const manifestKeys = [
-    event.rcloneResultsLarge.manifestKey,
-    event.rcloneResultsSmall.manifestKey,
-    event.rcloneResultsNeedThawSmall.manifestKey,
-    event.rcloneResultsNeedThawLarge.manifestKey,
-  ];
-
-  // Arrays to hold the result file keys from all manifests.
-  const pendingFiles: any[] = [];
-  const failedFiles: any[] = [];
-  const succeededFiles: any[] = [];
-
-  // Iterate over the manifest key produced by each map.
-  for (const manifestKey of manifestKeys) {
-    const getManifestCommand = new GetObjectCommand({
-      Bucket: event.workingBucket,
-      Key: manifestKey,
-    });
-
-    const getManifestResult = await client.send(getManifestCommand);
-    if (!getManifestResult.Body) {
-      throw new Error("Manifest S3 object Body is undefined");
-    }
-
-    const getManifestContent = await getManifestResult.Body.transformToString();
-
-    const manifest = JSON.parse(getManifestContent);
-
-    console.debug(JSON.stringify(manifest, null, 2));
-
-    const resultFiles = manifest["ResultFiles"];
-
-    if (!resultFiles)
-      throw new Error(
-        "AWS Steps Distributed map manifest.json is missing ResultFiles",
-      );
-
-    pendingFiles.push(...resultFiles["PENDING"]);
-    failedFiles.push(...resultFiles["FAILED"]);
-    succeededFiles.push(...resultFiles["SUCCEEDED"]);
-  }
-
-  // Validate manifest shape: all three status buckets must be arrays.
-  if (
-    !Array.isArray(pendingFiles) ||
-    !Array.isArray(failedFiles) ||
-    !Array.isArray(succeededFiles)
-  )
-    throw new Error(
-      "AWS Steps Distributed map manifest.json is missing an expected array for PENDING, FAILED or SUCCEEDED",
-    );
-
-  // Fail if PENDING.
-  // By the time this Lambda runs,
-  // the Distributed Map should have fully completed (no partial results).
-  if (pendingFiles.length > 0)
-    throw new Error(
-      "AWS Steps Distributed map manifest.json indicates there are PENDING results which is not a state we are expecting",
-    );
-
-  // Fail if FAILED.
-  // This Lambda is only expected to be run when the copy has fully succeeded.
-  // Maybe in future we want to handle partial success.
-  if (failedFiles.length > 0)
-    throw new Error("Copy is meant to succeed - but it had failed results");
-
-  // fileResults will hold the final collated results for all files.
-  const fileResults: Record<string, FileResult> = {};
-
-  // Process each SUCCEEDED result file to extract stats.
-  for (const succeededFile of succeededFiles) {
-    const getSuccessCommand = new GetObjectCommand({
-      Bucket: event.workingBucket,
-      Key: succeededFile["Key"],
-    });
-
-    const getSuccessResult = await client.send(getSuccessCommand);
-    if (!getSuccessResult.Body) {
-      throw new Error("Success S3 object Body is undefined");
-    }
-
-    const getSuccessContent = await getSuccessResult.Body.transformToString();
-
-    const content = getSuccessContent.trim();
-    const rows = content.startsWith("[")
-      ? JSON.parse(content)
-      : JSON.parse(
-          `[${content
-            .split(/\r?\n/)
-            .filter((l) => l.trim())
-            .join(",")}]`,
-        );
-
-    for (const row of rows) {
-      // NOTE/WARNING: this behaviour is very dependent on rclone and our interpretation
-      // of rclone stats - so if things start breaking this is where I would start
-      // looking
-
-      const source = row["source"];
-      // The name is the basename of the source..
-
-      // Original values
-      // const errors: number = rcloneRow["errors"];
-      // const lastError: number = rcloneRow["lastError"];
-      // const serverSideCopyBytes: number = rcloneRow["serverSideCopyBytes"];
-      // const elapsedTime = rcloneRow["elapsedTime"];
-      // const totalTransfers = rcloneRow["totalTransfers"];
-      // const retryError = rcloneRow["retryError"];
-
-      // Debbuged
-
-      // Is there were any errors during the transfer ?
-      const errors: number = row["errors"] ?? 0;
-
-      const name = basename(source);
-
-      // Compute the transfer speed
-      const elapsedSeconds = row["elapsedSeconds"] ?? 0;
-      const bytesTransferred = row["bytesTransferred"] ?? 0;
-      const speedMiBps =
-        elapsedSeconds > 0
-          ? bytesTransferred / elapsedSeconds / 1024 / 1024
-          : 0;
-
-      const destination = row["destination"];
-      const lastError: number = row["lastError"];
-      const retryError = row["retryError"]; // Not being passed
-
-      // firstly if we have been signalled an error - we need to report that
-      if (errors > 0) {
-        fileResults[name] = {
-          name: name,
-          status: "ERROR",
-          speed: speedMiBps,
-          message: lastError,
-          destination: destination,
-          bytesTransferred: bytesTransferred,
-          elapsedSeconds: elapsedSeconds,
-        };
-      } else {
-        // if we didn't end up transferring anything BUT there was no actual error AND
-        // we did a retry - then that probably means the source file didn't exist
-        if (bytesTransferred < 1 && retryError) {
-          fileResults[name] = {
-            name: name,
-            status: "ERROR",
-            speed: speedMiBps,
-            message: "source file did not exist so nothing was transferred",
-            destination: destination,
-            bytesTransferred: bytesTransferred,
-            elapsedSeconds: elapsedSeconds,
-          };
-        }
-        // if we didn't end up transferring anything BUT there was no actual error
-        // AND we didn't do any retries then changes are we skipped due to it already
-        // being at the destination
-        else if (bytesTransferred < 1 && !retryError) {
-          fileResults[name] = {
-            name: name,
-            status: "ALREADYCOPIED",
-            speed: speedMiBps,
-            message:
-              "destination file already exists with same checksum so nothing was transferred",
-            destination: destination,
-            bytesTransferred: bytesTransferred,
-            elapsedSeconds: elapsedSeconds,
-          };
-        } else {
-          fileResults[name] = {
-            name: name,
-            status: "COPIED",
-            speed: speedMiBps,
-            message: "",
-            destination: destination,
-            bytesTransferred: bytesTransferred,
-            elapsedSeconds: elapsedSeconds,
-          };
-        }
-      }
-    }
-  }
-
-  // debug results before we createting files
-  console.debug(JSON.stringify(fileResults, null, 2));
-
-  // --------------------------------------------------
-  // CVS ended copy generation and storage
-  // --------------------------------------------------
-  const output = stringify(Object.values(fileResults), {
-    header: true,
-    columns: {
-      name: "OBJECTNAME",
-      status: "TRANSFERSTATUS",
-      speed: "MBPERSEC",
-      message: "MESSAGE",
-      destination: "DESTINATION",
-      bytesTransferred: "BYTESTRANSFERRED",
-      elapsedSeconds: "ELAPSEDSECONDS",
-      // source: "SOURCE",
-      // copyMode: "COPYMODE",
-    },
-  });
-
-  // Write the CSV ended copy to the destination S3 bucket/folder
-  const csvKey = `${event.destinationPrefixKey}${event.destinationEndCopyRelativeKey}`;
-  const putCommand = new PutObjectCommand({
-    Bucket: event.destinationBucket,
-    Key: csvKey,
-    Body: output,
-  });
-
-  // --------------------------------------------------
-  // HTML ended copy report generation and storage
-  // --------------------------------------------------
-
   // Determine if we need to generate and store the HTML report(s)
   const includeReport = event.includeCopyReport;
   const retainReport = event.retainCopyReport;
+  const dryRun = event.dryRun;
 
-  if (includeReport || retainReport) {
-    const htmlReportName = "ENDED_COPY_REPORT.html";
+  if (dryRun) {
+    const summSmall = await readFileSummariesFromJsonl(
+      client,
+      event.inputCopySets.small.bucket,
+      event.inputCopySets.small.key,
+    );
+    const summLarge = await readFileSummariesFromJsonl(
+      client,
+      event.inputCopySets.large.bucket,
+      event.inputCopySets.large.key,
+    );
+    const summSmallThaw = await readFileSummariesFromJsonl(
+      client,
+      event.inputCopySets.smallThaw.bucket,
+      event.inputCopySets.smallThaw.key,
+    );
+    const summLargeThaw = await readFileSummariesFromJsonl(
+      client,
+      event.inputCopySets.largeThaw.bucket,
+      event.inputCopySets.largeThaw.key,
+    );
 
-    // Generate the HTML report
-    const html = createHtmlReport({
-      title: "Copy Results Report",
-      records: Object.values(fileResults) as FileResult[],
-      destinationBucket: event.destinationBucket,
-      destinationFolderKey: event.destinationPrefixKey,
-      costsSmall,
-      costsLarge,
-      costsSmallThaw,
-      costsLargeThaw,
-    });
-
-    // 1) Copy to the destination bucket/folder
-    if (includeReport) {
-      // TODO: define a better naming scheme for the HTML report (?)
-      const htmlKey = csvKey.replace("ENDED_COPY.csv", htmlReportName);
-
-      await client.send(
-        new PutObjectCommand({
-          Bucket: event.destinationBucket,
-          Key: htmlKey,
-          Body: html,
-          ContentType: "text/html; charset=utf-8",
-        }),
-      );
-    }
-
-    // 2) Extra copy to a specific S3 URI (sender retention)
     if (retainReport) {
+      const htmlReportName = "DRY_RUN_REPORT.html";
+
+      // Generate the HTML report
+      const html = createDryRunHtmlReport({
+        title: "Dry Run Report",
+        summSmall: summSmall,
+        summLarge: summLarge,
+        summSmallThaw: summSmallThaw,
+        summLargeThaw: summLargeThaw,
+      });
+
       const sourceFilePrefix = dirname(event.copyInstructionsKey) + "/";
       const retainReportKey = sourceFilePrefix + htmlReportName;
 
@@ -413,7 +223,287 @@ export async function handler(event: InvokeEvent) {
         }),
       );
     }
+    return;
+  } else {
+    //  This is the actual COPY report
+
+    // Each Distributed Map we run (Large, Small, NeedThawSmall, NeedThawLarge) writes a
+    // `manifest.json` to S3. The manifest is generated by Step Functions and summarises the
+    // per-item result files for that map run (grouped into PENDING / FAILED / SUCCEEDED).
+    //
+    // Example:
+    // {"DestinationBucket":"elsa-data-tmp",
+    // "MapRunArn":"arn:aws:states:ap-southeast-2:12345678:mapRun:CopyOutStateMachineABCD/4474d22f-4056-30e3-978c-027016edac90:0c17ffd6-e8ad-44c0-a65b-a8b721007241",
+    // "ResultFiles":{
+    //     "FAILED":[],
+    //     "PENDING":[],
+    //     "SUCCEEDED":[{"Key":"copy-out-test-working/a6faea86c066cd90/1-objects-to-copy.tsv/0c17ffd6-e8ad-44c0-a65b-a8b721007241/SUCCEEDED_0.json",
+    //                   "Size":2887}]}}
+
+    // We read and merge the manifests from all four maps to get the complete set of results.
+
+    const manifestKeys = [
+      event.rcloneResultsLarge.manifestKey,
+      event.rcloneResultsSmall.manifestKey,
+      event.rcloneResultsNeedThawSmall.manifestKey,
+      event.rcloneResultsNeedThawLarge.manifestKey,
+    ];
+
+    // Arrays to hold the result file keys from all manifests.
+    const pendingFiles: any[] = [];
+    const failedFiles: any[] = [];
+    const succeededFiles: any[] = [];
+
+    // Iterate over the manifest key produced by each map.
+    for (const manifestKey of manifestKeys) {
+      const getManifestCommand = new GetObjectCommand({
+        Bucket: event.workingBucket,
+        Key: manifestKey,
+      });
+
+      const getManifestResult = await client.send(getManifestCommand);
+      if (!getManifestResult.Body) {
+        throw new Error("Manifest S3 object Body is undefined");
+      }
+
+      const getManifestContent =
+        await getManifestResult.Body.transformToString();
+
+      const manifest = JSON.parse(getManifestContent);
+
+      console.debug(JSON.stringify(manifest, null, 2));
+
+      const resultFiles = manifest["ResultFiles"];
+
+      if (!resultFiles)
+        throw new Error(
+          "AWS Steps Distributed map manifest.json is missing ResultFiles",
+        );
+
+      pendingFiles.push(...resultFiles["PENDING"]);
+      failedFiles.push(...resultFiles["FAILED"]);
+      succeededFiles.push(...resultFiles["SUCCEEDED"]);
+    }
+
+    // Validate manifest shape: all three status buckets must be arrays.
+    if (
+      !Array.isArray(pendingFiles) ||
+      !Array.isArray(failedFiles) ||
+      !Array.isArray(succeededFiles)
+    )
+      throw new Error(
+        "AWS Steps Distributed map manifest.json is missing an expected array for PENDING, FAILED or SUCCEEDED",
+      );
+
+    // Fail if PENDING.
+    // By the time this Lambda runs,
+    // the Distributed Map should have fully completed (no partial results).
+    if (pendingFiles.length > 0)
+      throw new Error(
+        "AWS Steps Distributed map manifest.json indicates there are PENDING results which is not a state we are expecting",
+      );
+
+    // Fail if FAILED.
+    // This Lambda is only expected to be run when the copy has fully succeeded.
+    // Maybe in future we want to handle partial success.
+    if (failedFiles.length > 0)
+      throw new Error("Copy is meant to succeed - but it had failed results");
+
+    // fileResults will hold the final collated results for all files.
+    const fileResults: Record<string, FileResult> = {};
+
+    // Process each SUCCEEDED result file to extract stats.
+    for (const succeededFile of succeededFiles) {
+      const getSuccessCommand = new GetObjectCommand({
+        Bucket: event.workingBucket,
+        Key: succeededFile["Key"],
+      });
+
+      const getSuccessResult = await client.send(getSuccessCommand);
+      if (!getSuccessResult.Body) {
+        throw new Error("Success S3 object Body is undefined");
+      }
+
+      const getSuccessContent = await getSuccessResult.Body.transformToString();
+
+      const content = getSuccessContent.trim();
+      const rows = content.startsWith("[")
+        ? JSON.parse(content)
+        : JSON.parse(
+            `[${content
+              .split(/\r?\n/)
+              .filter((l) => l.trim())
+              .join(",")}]`,
+          );
+
+      for (const row of rows) {
+        // NOTE/WARNING: this behaviour is very dependent on rclone and our interpretation
+        // of rclone stats - so if things start breaking this is where I would start
+        // looking
+
+        const source = row["source"];
+        // The name is the basename of the source..
+
+        // Original values
+        // const errors: number = rcloneRow["errors"];
+        // const lastError: number = rcloneRow["lastError"];
+        // const serverSideCopyBytes: number = rcloneRow["serverSideCopyBytes"];
+        // const elapsedTime = rcloneRow["elapsedTime"];
+        // const totalTransfers = rcloneRow["totalTransfers"];
+        // const retryError = rcloneRow["retryError"];
+
+        // Debbuged
+
+        // Is there were any errors during the transfer ?
+        const errors: number = row["errors"] ?? 0;
+
+        const name = basename(source);
+
+        // Compute the transfer speed
+        const elapsedSeconds = row["elapsedSeconds"] ?? 0;
+        const bytesTransferred = row["bytesTransferred"] ?? 0;
+        const speedMiBps =
+          elapsedSeconds > 0
+            ? bytesTransferred / elapsedSeconds / 1024 / 1024
+            : 0;
+
+        const destination = row["destination"];
+        const lastError: number = row["lastError"];
+        const retryError = row["retryError"]; // Not being passed
+
+        // firstly if we have been signalled an error - we need to report that
+        if (errors > 0) {
+          fileResults[name] = {
+            name: name,
+            status: "ERROR",
+            speed: speedMiBps,
+            message: lastError,
+            destination: destination,
+            bytesTransferred: bytesTransferred,
+            elapsedSeconds: elapsedSeconds,
+          };
+        } else {
+          // if we didn't end up transferring anything BUT there was no actual error AND
+          // we did a retry - then that probably means the source file didn't exist
+          if (bytesTransferred < 1 && retryError) {
+            fileResults[name] = {
+              name: name,
+              status: "ERROR",
+              speed: speedMiBps,
+              message: "source file did not exist so nothing was transferred",
+              destination: destination,
+              bytesTransferred: bytesTransferred,
+              elapsedSeconds: elapsedSeconds,
+            };
+          }
+          // if we didn't end up transferring anything BUT there was no actual error
+          // AND we didn't do any retries then changes are we skipped due to it already
+          // being at the destination
+          else if (bytesTransferred < 1 && !retryError) {
+            fileResults[name] = {
+              name: name,
+              status: "ALREADYCOPIED",
+              speed: speedMiBps,
+              message:
+                "destination file already exists with same checksum so nothing was transferred",
+              destination: destination,
+              bytesTransferred: bytesTransferred,
+              elapsedSeconds: elapsedSeconds,
+            };
+          } else {
+            fileResults[name] = {
+              name: name,
+              status: "COPIED",
+              speed: speedMiBps,
+              message: "",
+              destination: destination,
+              bytesTransferred: bytesTransferred,
+              elapsedSeconds: elapsedSeconds,
+            };
+          }
+        }
+      }
+    }
+
+    // debug results before we createting files
+    console.debug(JSON.stringify(fileResults, null, 2));
+
+    // --------------------------------------------------
+    // CVS ended copy generation and storage
+    // --------------------------------------------------
+    const output = stringify(Object.values(fileResults), {
+      header: true,
+      columns: {
+        name: "OBJECTNAME",
+        status: "TRANSFERSTATUS",
+        speed: "MBPERSEC",
+        message: "MESSAGE",
+        destination: "DESTINATION",
+        bytesTransferred: "BYTESTRANSFERRED",
+        elapsedSeconds: "ELAPSEDSECONDS",
+        // source: "SOURCE",
+        // copyMode: "COPYMODE",
+      },
+    });
+
+    // Write the CSV ended copy to the destination S3 bucket/folder
+    const csvKey = `${event.destinationPrefixKey}${event.destinationEndCopyRelativeKey}`;
+    const putCommand = new PutObjectCommand({
+      Bucket: event.destinationBucket,
+      Key: csvKey,
+      Body: output,
+    });
+
+    // --------------------------------------------------
+    // HTML ended copy report generation and storage
+    // --------------------------------------------------
+
+    if (includeReport || retainReport) {
+      const htmlReportName = "ENDED_COPY_REPORT.html";
+
+      // Generate the HTML report
+      const html = createHtmlReport({
+        title: "Copy Results Report",
+        records: Object.values(fileResults) as FileResult[],
+        destinationBucket: event.destinationBucket,
+        destinationFolderKey: event.destinationPrefixKey,
+        costsSmall,
+        costsLarge,
+        costsSmallThaw,
+        costsLargeThaw,
+      });
+
+      // 1) Copy to the destination bucket/folder
+      if (includeReport) {
+        // TODO: define a better naming scheme for the HTML report (?)
+        const htmlKey = csvKey.replace("ENDED_COPY.csv", htmlReportName);
+
+        await client.send(
+          new PutObjectCommand({
+            Bucket: event.destinationBucket,
+            Key: htmlKey,
+            Body: html,
+            ContentType: "text/html; charset=utf-8",
+          }),
+        );
+      }
+
+      // 2) Extra copy to a specific S3 URI (sender retention)
+      if (retainReport) {
+        const sourceFilePrefix = dirname(event.copyInstructionsKey) + "/";
+        const retainReportKey = sourceFilePrefix + htmlReportName;
+
+        await client.send(
+          new PutObjectCommand({
+            Bucket: event.workingBucket,
+            Key: retainReportKey,
+            Body: html,
+            ContentType: "text/html; charset=utf-8",
+          }),
+        );
+      }
+    }
+    await client.send(putCommand);
+    return output;
   }
-  await client.send(putCommand);
-  return output;
 }
