--- packages/steps-s3-copy/lambda/summarise-dryrun-lambda/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "summarise-copy-lambda",
-  "version": "1.0.0",
-  "main": "summarise-copy-lambda.ts",
-  "dependencies": {
-    "@aws-sdk/client-s3": "3.705.0",
-    "@types/aws-lambda": "8.10.93",
-    "csv-stringify": "6.5.2"
-  }
-}
